Unify the embassy_net bootstrapping code for all esp examples
